pyproject.toml
[project]
name = "inbox-triage"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27,<1",
    "scikit-learn>=1.5,<2",
    "joblib>=1.4,<2",
    "click>=8.1,<9",
    "rich>=13.9,<14",
    "python-dotenv>=1.0,<2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
inbox-triage = "inbox_triage.cli:cli"
